+++ #633 - Server/src/com/rs/game/player/actions/Rest.java
@@ -25,9 +25,9 @@ public class Rest extends Action {
 		if (!process(player))
 			return false;
 		index = Utils.random(REST_DEFS.length);
-		player.setResting(musician ? 2 : 1);
+		player.setResting((byte) (musician ? 2 : 1));
 		player.setNextAnimation(new Animation(REST_DEFS[index][0]));
-		player.getAppearence().setRenderEmote(REST_DEFS[index][1]);
+		player.getAppearence().setRenderEmote((short) REST_DEFS[index][1]);
 		return true;
 	}
 
@@ -53,10 +53,10 @@ public class Rest extends Action {
 
 	@Override
 	public void stop(Player player) {
-		player.setResting(0);
+		player.setResting((byte) 0);
 		player.setNextAnimation(new Animation(REST_DEFS[index][2]));
 		player.getEmotesManager().setNextEmoteEnd();
-		player.getAppearence().setRenderEmote(-1);
+		player.getAppearence().setRenderEmote((short) -1);
 	}
 
 }
